+++ client/src/components/Filtros/Filtros.jsx
@@ -7,21 +7,6 @@ import styles from './Filtros.module.css'
 import { getAllProducts } from "../../redux/Actions/getAllProducts";
 
 export default function Filtros() {
-/*   const dispatch = useDispatch();
-  const productos = useSelector((state) => state.productos);
-  const categorias = useSelector((state) => state.categorias);
-  const toShow = useSelector((state) => state.toShow);
-
-  const [categoria, setCategoria] = useState("Todas");
-  const [order, setOrder] = useState("Ninguno");
-  const [precio, setPrecio] = useState("Todos"); */
-  /* const [fecha, setFecha] = useState('Ninguna') */
-  // const dispatch = useDispatch();
-  // useEffect(()=>{
-  //   return ()=>{
-  //     dispatch(getAllProducts())
-  //   }
-  // })
 
   const [orderName, setOrderName] = useState('DEFAULT')
   const [orderPrecio, setOrderPrecio] = useState('DEFAULT')
@@ -29,15 +14,15 @@ export default function Filtros() {
   return (
     <>
       <div className={styles.divMain}>
-        <div>
+        <div className={styles.divOne}>
           <FiltroCategoria orderName={orderName} setOrderName={setOrderName}
           orderPrecio={orderPrecio} setOrderPrecio={setOrderPrecio}/>
         </div>
-        <div>
+        <div className={styles.divTwo}>
           <OrderName orderName={orderName} setOrderName={setOrderName}
           orderPrecio={orderPrecio} setOrderPrecio={setOrderPrecio}/>
         </div>
-        <div>
+        <div className={styles.divThree}>
           <OrderPrice orderName={orderName} setOrderName={setOrderName}
            orderPrecio={orderPrecio} setOrderPrecio={setOrderPrecio}/>
         </div>
