+++ client/src/paginas/Catalogo/Catalogo.jsx
@@ -6,6 +6,7 @@ import { getAllProducts } from "../../redux/Actions/getAllProducts";
 import style from "./Catalogo.module.css";
 import CardsArr from "../../components/Cards/CardsArr";
 import Footer from "../../components/Footer/Footer";
+import BotonSwitch from "../../elementos/BotonSwitch";
 
 const Catalogo = () => {
   const dispatch = useDispatch();
@@ -23,26 +24,18 @@ const Catalogo = () => {
     allProducts?.length === 0 && dispatch(getAllProducts());
   }, [allProducts, dispatch]);
 
-  const handleCheckboxChange = (event) => {
-    const { checked } = event.target;
-    if(checked===false){
-      dispatch(getAllProducts())
-    }
-    setFiltroActivo(!filtroActivo);
-  };
+
 
   return (
     <>
-      <div>
-          <h3>Filtros</h3>
-          <input
-            className={style.active}
-            type="checkbox"
-            checked={filtroActivo}
-            onChange={(event)=>handleCheckboxChange(event)}
-          />
-          {filtroActivo && <Filtros />}
+      <div className={style.divFiltros}>
+        
+        <div className={style.divOne}><h4>Filtros</h4><BotonSwitch  filtroActivo={filtroActivo} setFiltroActivo={setFiltroActivo} /></div>
+
+        <div className={style.divTwo}>{filtroActivo && <Filtros />}</div>
+            
       </div>
+
       <div>
         {filtroActivo ? (
           <>
