+++ client/src/paginas/Catalogo/Catalogo.module.css
@@ -1,6 +1,23 @@
 .filtros{
     display: flex;
+    height: 50px;
+}
+.divFiltros{
+    display: flex;
     justify-content: center;
-    gap: 80px;
-    line-height: 70px;
+    align-items: center;
+    height: 150px;
+    width: 100%;
+}
+
+.divOne{
+    display: flex;
+    width: 25%;
+    justify-content: center;
+    align-items: center;
+    font-family: 'Gotham',sans-serif;
+}
+
+.divTwo{
+    width: 75%;
 }
